# Backend/docs/swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  NotificationsCount:
    description: NotificationsCount represents the count of notifications.
    properties:
      count:
        description: Count is the number of notifications.
        type: integer
    type: object
  NotificationsDetails:
    description: NotificationsDetails represents detailed information about a specific
      notification.
    properties:
      content:
        description: Content is the textual content of the notification.
        type: string
      createdAt:
        description: CreatedAt is the timestamp indicating when the notification was
          created.
        type: string
      id:
        description: ID is the unique identifier for the notification.
        type: string
      seen:
        description: Seen is a boolean indicating whether the notification has been
          seen or not.
        type: boolean
      type:
        description: Type is the type or category of the notification.
        type: string
    type: object
  NotificationsIn:
    description: NotificationsIn represents the input structure for updating the "seen"
      status of notifications.
    properties:
      seen:
        description: Seen is a boolean indicating whether the notification has been
          seen or not. It is required.
        type: boolean
    required:
    - seen
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /notifications/{userID}:
    get:
      description: Get all notifications.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/NotificationsDetails'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notifications
      tags:
      - Notifications
  /notifications/{userID}/{ID}:
    delete:
      consumes:
      - application/json
      description: Delete one notification.
      parameters:
      - description: Company ID
        in: path
        name: userID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete notification
      tags:
      - Notifications
    get:
      description: Get one notification.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/NotificationsDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notification
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Update one notification.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Notification ID
        in: path
        name: ID
        required: true
        type: string
      - description: Notification query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/NotificationsIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update notification
      tags:
      - Notifications
  /notifications/{userID}/count:
    get:
      description: Get all notifications count.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/NotificationsCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notifications count
      tags:
      - Notifications
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"